+++ gemset.nix
@@ -81,7 +81,7 @@
       path = ./.;
       type = "path";
     };
-    version = "5.0.2";
+    version = "5.0.3";
   };
   cgi = {
     groups = ["completions" "default"];